requirements.txt
djangorestframework
python-dotenv
requests